Added ADC driver, modified java ADC API. Added driver for BH1620FVC light sensor. Removed printf on opt functions. Java Sensors API up and running.

+++ VirtualSense/DJ_VirtualMachine.1.0/darjeeling.1.1/src/lib/virtualsense/javax/virtualsense/sensors/Light.java
@@ -32,7 +32,6 @@ public class Light
 	 * Reads light value from the BH1620FVC sensor.	 
 	 * @return measured light express in mLx.
 	 */
-	//TODO: c'è un motivo per ritornare un int e non uno short ?
-	public static native int getValue();
+	public static native short getValue();
 
 }
